pyproject.toml
[project]
name = "self-assessment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.3",
    "langchain-anthropic>=1.0.1",
    "langchain-google-genai>=3.0.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.1",
    "python-dotenv>=1.2.1",
]
